verify-python/requirements.txt
requests==2.21.0
pycryptodome==3.7.3